refactor: alert types
