+++ projects/ion/src/lib/alert/alert.component.spec.ts
@@ -1,8 +1,9 @@
 import { CommonModule } from '@angular/common';
 import { fireEvent, render, screen } from '@testing-library/angular';
 import { StatusType } from '../core/types/status';
-import { IonAlertComponent, IonAlertProps } from './alert.component';
+import { IonAlertComponent } from './alert.component';
 import { IonIconModule } from '../icon/icon.module';
+import { IonAlertProps } from '../core/types/alert';
 
 const defaultValue: IonAlertProps = {
   message: 'Mensagem padrão',
