+++ projects/ion/src/lib/core/types/alert.ts
@@ -1,0 +1,8 @@
+import { StatusType } from './status';
+
+export interface IonAlertProps {
+  message: string;
+  type?: StatusType;
+  closable?: boolean;
+  hideBackground?: boolean;
+}
