+++ projects/ion/src/lib/core/types/index.ts
@@ -4,3 +4,4 @@ export * from './status';
 export * from './tooltip';
 export * from './card';
 export * from './tab-group';
+export * from './alert';
